+++ Core/Util/Configs/Components/ConfigAudio.cs
@@ -33,9 +33,9 @@ public class ConfigAudio: ConfigElement<ConfigAudio>
     [OptionMenu(OptionSectionType.Audio, "Same Sound Limit")]
     public readonly ConfigValue<int> SameSoundLimit = new(0, GreaterOrEqual(0));
 
-    [ConfigInfo("Limit same sounds window in ticks. 0 = off.")]
+    [ConfigInfo("Limit same sounds window in ticks.")]
     [OptionMenu(OptionSectionType.Audio, "Same Sound Window")]
-    public readonly ConfigValue<int> SameSoundWindow = new(1, GreaterOrEqual(0));
+    public readonly ConfigValue<int> SameSoundWindow = new(1, GreaterOrEqual(1));
 
     [ConfigInfo("Randomize sound pitch.")]
     [OptionMenu(OptionSectionType.Audio, "Randomize Pitch", spacer: true)]